# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/listings": {
      "get": {
        "description": "Returns published listings. With mine=true it returns the caller’s own listings in any status; moderators may filter by any status.",
        "operationId": "Listings.search",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "hy",
                "ru",
                "en"
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "published_desc",
              "type": "string",
              "enum": [
                "published_desc",
                "price_asc",
                "price_desc",
                "price_per_sqm_asc",
                "price_per_sqm_desc",
                "area_asc",
                "area_desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "PUBLISHED",
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_REVIEW",
                "PUBLISHED",
                "REJECTED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "mine",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "priceMin",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "priceMax",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "pricePerSqmMax",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "roomsMin",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10,
              "type": "integer"
            }
          },
          {
            "name": "roomsMax",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 10,
              "type": "integer"
            }
          },
          {
            "name": "areaMin",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "areaMax",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "floorMin",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "floorMax",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "ceilingMin",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 2,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "yearMin",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1850,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "yearMax",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1850,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "districts",
            "required": false,
            "in": "query",
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "buildingTypes",
            "required": false,
            "in": "query",
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "conditions",
            "required": false,
            "in": "query",
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "heating",
            "required": false,
            "in": "query",
            "schema": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          {
            "name": "excludeGroundFloor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "excludeTopFloor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "hasElevator",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "hasParking",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "hasBalcony",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "hasStorage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "seismicRetrofit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "docsVerified",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "priceNegotiable",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "nearLat",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "type": "number"
            }
          },
          {
            "name": "nearLon",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "type": "number"
            }
          },
          {
            "name": "radiusM",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 100,
              "maximum": 20000,
              "type": "integer"
            }
          },
          {
            "name": "bbox",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingsPageDto"
                }
              }
            }
          }
        },
        "summary": "Structured search with filters, sorting and cursor pagination",
        "tags": [
          "listings"
        ]
      },
      "post": {
        "description": "Any signed-in account may create one. A regular user’s listing enters the review queue; a verified agent’s is published immediately.",
        "operationId": "Listings.create",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Response locale for listing texts; falls back to the account preference, then Accept-Language, then hy",
            "schema": {
              "enum": [
                "hy",
                "ru",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetailDto"
                }
              }
            }
          },
          "409": {
            "description": "The account already holds its maximum of live listings"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a listing for an existing building",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/listings/{idOrPublicId}": {
      "get": {
        "operationId": "Listings.getOne",
        "parameters": [
          {
            "name": "idOrPublicId",
            "required": true,
            "in": "path",
            "schema": {
              "oneOf": [
                {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                {
                  "pattern": "^L-[A-HJ-NP-Z2-9]{6}$",
                  "type": "string"
                }
              ]
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Response locale for listing texts; falls back to the account preference, then Accept-Language, then hy",
            "schema": {
              "enum": [
                "hy",
                "ru",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Unknown, or not visible to the caller"
          }
        },
        "summary": "Listing detail by id or public id",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/listings/{id}": {
      "patch": {
        "description": "Status is not editable here; use the transitions endpoint.",
        "operationId": "Listings.update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Response locale for listing texts; falls back to the account preference, then Accept-Language, then hy",
            "schema": {
              "enum": [
                "hy",
                "ru",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "Not the owner"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update attributes or translations of a listing",
        "tags": [
          "listings"
        ]
      },
      "delete": {
        "operationId": "Listings.archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "Not the owner and not a moderator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a listing (soft delete)",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/listings/{id}/transitions": {
      "post": {
        "description": "SUBMIT, PUBLISH, APPROVE, REJECT, REVISE and ARCHIVE. A transition that is not legal from the current status is a conflict, not a bad request.",
        "operationId": "Listings.transition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Response locale for listing texts; falls back to the account preference, then Accept-Language, then hy",
            "schema": {
              "enum": [
                "hy",
                "ru",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingTransitionBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "The caller may not perform this transition"
          },
          "409": {
            "description": "Illegal from the listing’s current status"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Move a listing through the moderation lifecycle",
        "tags": [
          "listings"
        ]
      }
    },
    "/api/listings/{id}/permanent": {
      "delete": {
        "description": "Administrators only. Archiving is the reversible option.",
        "operationId": "Listings.destroy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "Requires ADMIN"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a listing and all of its history, permanently",
        "tags": [
          "listings"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartEstate API",
    "description": "AI-assisted apartment search and buying advisor for the Armenian market. All request and response shapes are defined once as Zod schemas in @smartestate/contracts; this document is generated from them.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "smartestate_refresh"
      }
    },
    "schemas": {
...
      "ListingsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "publicId": {
                  "type": "string",
                  "pattern": "^L-[A-HJ-NP-Z2-9]{6}$"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "DRAFT",
                    "PENDING_REVIEW",
                    "PUBLISHED",
                    "REJECTED",
                    "ARCHIVED"
                  ]
                },
                "locale": {
                  "type": "string",
                  "enum": [
                    "hy",
                    "ru",
                    "en"
                  ]
                },
                "title": {
                  "type": "string"
                },
                "priceAmd": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "pricePerSqmAmd": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "originalCurrency": {
                  "type": "string",
                  "enum": [
                    "AMD",
                    "USD",
                    "EUR"
                  ]
                },
                "originalPrice": {
                  "nullable": true,
                  "type": "number"
                },
                "priceNegotiable": {
                  "type": "boolean"
                },
                "rooms": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "totalArea": {
                  "type": "number"
                },
                "floor": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "totalFloors": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "buildingType": {
                  "type": "string",
                  "enum": [
                    "STONE",
                    "PANEL",
                    "MONOLITH",
                    "KHRUSHCHYOVKA",
                    "STALINKA",
                    "NEW_BUILD"
                  ]
                },
                "condition": {
                  "type": "string",
                  "enum": [
                    "NEEDS_REPAIR",
                    "OLD_RENOVATION",
                    "GOOD",
                    "EURO_RENOVATION",
                    "DESIGNER"
                  ]
                },
                "district": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9-]{1,40}$"
                    },
                    "name": {
                      "type": "object",
                      "properties": {
                        "hy": {
                          "type": "string"
                        },
                        "ru": {
                          "type": "string"
                        },
                        "en": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "hy",
                        "ru",
                        "en"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "slug",
                    "name"
                  ],
                  "additionalProperties": false
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "lat": {
                      "type": "number",
                      "minimum": -90,
                      "maximum": 90
                    },
                    "lon": {
                      "type": "number",
                      "minimum": -180,
                      "maximum": 180
                    }
                  },
                  "required": [
                    "lat",
                    "lon"
                  ],
                  "additionalProperties": false
                },
                "thumbnailUrl": {
                  "nullable": true,
                  "type": "string",
                  "format": "uri"
                },
                "publishedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                }
              },
              "required": [
                "id",
                "publicId",
                "status",
                "locale",
                "title",
                "priceAmd",
                "pricePerSqmAmd",
                "originalCurrency",
                "originalPrice",
                "priceNegotiable",
                "rooms",
                "totalArea",
                "floor",
                "totalFloors",
                "buildingType",
                "condition",
                "district",
                "location",
                "thumbnailUrl",
                "publishedAt"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
      "ListingTransitionBodyDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "SUBMIT",
              "PUBLISH",
              "APPROVE",
              "REJECT",
              "REVISE",
              "ARCHIVE"
            ]
          },
          "reason": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000
          }
        },
        "required": [
          "action"
        ]
      },
...
    }
  }
}